Move mute logic to TrackMiddleware/MuteLocal.ts
