+++ services/conference/src/scripts/models/TrackMiddleware/index.ts
@@ -1,0 +1,1 @@
+import './MuteLocal'
